refactor(components): moved nested components into base components directory
